+++ ThirdPartyAdapters/tapjoy/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Tapjoy Adapter for Google Mobile Ads SDK for Android Changelog
 
-## Unreleased
+## 11.11.0.0
+- Updated the adapter to make it compatible with Tapjoy SDK version 11.11.0.
+
+## 11.10.2.0
 - Updated the default reward amount to 1.
+- Verified compatibility with Tapjoy SDK version 11.10.2.
 
 ## 11.10.1.0
 - Verified compatibility with Tapjoy SDK version 11.10.1.